+++ lib/src/theme/src/light_theme.dart
@@ -48,8 +48,6 @@ ThemeData generateLightTheme({
   /// ```
   AppFont? bodyFont,
 }) {
-  bool isIOS = !kIsWeb && Platform.isIOS;
-
   MaterialColor color = getThemeColor(
     theme: theme,
     color: mainColor,